requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
annotated-types==0.7.0
astunparse==1.6.3
audioread==3.0.1
black==25.1.0
cachetools==5.5.2
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
ffmpeg-python==0.2.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
gast==0.6.0
google-auth==2.39.0
google-genai==1.10.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
html5tagger==1.3.0
httptools==0.6.4
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.7
matplotlib==3.10.0
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
pooch==1.8.2
protobuf==5.29.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.1
python-dotenv==1.1.0
pytz==2025.1
rsa==4.9
sanic==25.3.0
sanic-routing==23.12.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.12.0
tensorflowjs==3.18.0
termcolor==2.5.0
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
tracerite==1.1.1
triton==3.2.0
typing-inspection==0.4.0
tzdata==2025.1
ujson==5.10.0
ultralytics==8.3.75
ultralytics-thop==2.0.14
uvloop==0.21.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
